Add Pod Name and Namespace
The protobuf in nimbess-agent has changed to allow for the pod name and
namespace to be passed in the CNI request. This is required in order to
annotate the pod with the Fast Path information. This change parses the
pod name and namespace to pass along with the CNI request.

Additionally, dependency managemnt is moved to go modules.

Signed-off-by: Tim Rozet <trozet@redhat.com>